+++ changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.8.0](https://github.com/arianrhodsandlot/nostalgist/compare/v0.7.0...v0.8.0) - 2024-02-07
 ### Changed
 - update default retroarch-emscripten-build version to v1.17.0
 - disable some default inputs in RetroArch